rename component folder
