+++ web/src/app/admin/admin-message-logs/DebugMessageCard.tsx
